moved logic to add fake item from MenuFragment.kt to ExposureInformationRepository.kt
