+++ app/src/main/java/org/covidwatch/android/data/exposureinformation/ExposureInformationRepository.kt
@@ -1,10 +1,20 @@
 package org.covidwatch.android.data.exposureinformation
 
+import android.app.Activity
+import android.content.Context
 import androidx.lifecycle.LiveData
+import androidx.navigation.Navigation.findNavController
+import androidx.navigation.fragment.findNavController
+import kotlinx.coroutines.GlobalScope
+import org.covidwatch.android.R
 import org.covidwatch.android.data.CovidExposureInformation
 import org.covidwatch.android.data.toCovidExposureInformation
 import org.covidwatch.android.exposurenotification.RandomEnObjects
+import org.covidwatch.android.exposurenotification.TestExposureNotification
+import org.covidwatch.android.extension.io
 import org.covidwatch.android.extension.mutableLiveData
+import androidx.navigation.fragment.findNavController
+
 
 class ExposureInformationRepository(private val local: ExposureInformationLocalSource) {
     suspend fun saveExposureInformation(exposureInformation: List<CovidExposureInformation>) {
@@ -26,6 +36,32 @@ class ExposureInformationRepository(private val local: ExposureInformationLocalS
 
     fun randomExposureInformation(): List<CovidExposureInformation> {
         return local.randomExposureInformation()
+    }
+
+    fun addFakeItem(context: Context){
+        var returnExposureInformationList : List<CovidExposureInformation>
+        val testExposureNotification = TestExposureNotification()
+        val covidExposureInformation: CovidExposureInformation =
+            RandomEnObjects.exposureInformation.toCovidExposureInformation()
+        var exposureInformationList: List<CovidExposureInformation> =
+            listOf(covidExposureInformation)
+
+        GlobalScope.io {
+            returnExposureInformationList = saveOneGetAll(exposureInformationList)
+            //sum up risk exposures from returnExposureInformationList and pass to TestExposureNotification
+            testExposureNotification.saveExposureSummaryInPreferences(context,covidExposureInformation,returnExposureInformationList.size)
+        }
+    }
+
+    //Save the new exposureInformation object to the database
+    //Read all the exposureInformation objects from the database into a list
+    suspend private fun saveOneGetAll(
+        exposureInformationList: List<CovidExposureInformation>): List<CovidExposureInformation>
+    {
+        saveExposureInformation(exposureInformationList)
+        var newExposureInformationList: List<CovidExposureInformation>
+        newExposureInformationList = randomExposureInformation()
+        return newExposureInformationList
     }
 
 }
